Remove separate getters/setters for neighbouring tiles (#21)
* Remove individual neighbor tile setters/getters

get_tile_in_direction() and set_adjacent_tile() can do the work of all the setters/getters.

* Refactor for new implementation of `get_tile_in_direction()`

`get_tile_in_direction()` now has the same implementation as the individual getters.

Check for instance of `NullModel` in the Move event and controller instead of catching the tile exception since now, an exception is only thrown if the tile is `None` (not `NullModel`).

`NullModel`s are set on purpose anyways so it's not an issue to have a `NullModel`. It's only an issue if the tile is `None`.